% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{pivot}
\alias{pivot}
\title{PIVOT laucher}
\usage{
pivot(args = "launcher")
}
\description{
Serves as a gateway for all analysis modules and allow users to monitor the R session state and clean when needed.
}
\examples{
pivot()
}
